pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.5",
    "crawl4ai>=0.7.0",
    "qdrant-client>=1.14.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "neo4j>=5.28.1",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.1",
    "numpy>=1.24.0",
    "gitpython>=3.1.44",
    "psutil>=5.9.0",
]

[dependency-groups]
dev = [
    "watchdog",
    "watchfiles>=1.1.0",
]
